+++ examples/how_long_until.rs
@@ -7,7 +7,7 @@ extern crate dont_disappear;
 extern crate read_input;
 use chrono::offset::{Local, TimeZone};
 use chrono::prelude::*;
-use read_input::*;
+use read_input::prelude::*;
 use std::str::FromStr;
 
 struct DateDDMMYY(DateTime<Local>);