+++ examples/point_input.rs
@@ -6,7 +6,7 @@
 extern crate dont_disappear;
 extern crate read_input;
 
-use read_input::*;
+use read_input::prelude::*;
 use std::str::FromStr;
 
 #[derive(Debug)]