+++ src/prelude.rs
@@ -1,0 +1,4 @@
+pub use input_new;
+pub use shortcut::simple_input;
+pub use shortcut::valid_input;
+pub use shortcut::with_description;