+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 ### New features
 - New `InputBuilderOnce` `struct` to make the builder more flexible.
 - Make methods take `impl ToString` rather than `&str` to improve flexibility.
+- New `input_new_d()` function.
 ### **Breaking changes**
 - Module restructure.
+- Move to `edition = "2018"`
 
 # 0.6.1 and 0.6.2
 ### Documentation updates