+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# __Next__
+# __0.7.0__
 ### New features
 - New `InputBuilderOnce` `struct` to make the builder more flexible.
 - Make methods take `impl ToString` rather than `&str` to improve flexibility.